Handle corrupted tgz file issue while harvesting

+++ H2TauTau/python/harvest/harvest.py
@@ -32,7 +32,9 @@ def harvest_one(info, destination, ntgzs=None):
      print('there')
      tmpdir = tempfile.mkdtemp()
      fetch(info, tmpdir, ntgzs)
-     unpack(info, tmpdir, ntgzs)
+     unpack_exit_code = unpack(info, tmpdir, ntgzs)
+     if not unpack_exit_code == 0:
+          return info 
      hadd(tmpdir)
      tmpsampledir = '/'.join([tmpdir,info['name']])
      destsampledir = '/'.join([destination, info['name']])
@@ -177,15 +179,24 @@ def unpack(info, destination, ntgzs=None):
                ntgzs = len(tgzs)
           for tgz in tgzs[:ntgzs]: 
                print 'unpacking', tgz
-               os.system('tar -zxf {}'.format(tgz))
+               untar_exit_code = os.system('tar -zxf {}'.format(tgz))
                # tgz is e.g. heppyOutput_43.tgz
                # so index is 43
                index = int(os.path.splitext(tgz)[0].split('_')[1]) + subdid
-               chunkname = '{}_Chunk{}'.format(info['name'], index)
-               chunks[subd].append(chunkname)
-               os.rename('Output', '../'+chunkname)
-               os.remove(tgz)
+               if not untar_exit_code == 0:
+                    os.system("echo 'Problem with sample {}, chunk {}, version {}, submitted on {}' >> $CMSSW_BASE/src/CMGTools/H2TauTau/scripts/tgzs_problems.log".format(
+                              info['sample'],
+                              index,
+                              info['sample_version'],
+                              info['sub_date']))
+                    return untar_exit_code
+               else:
+                    chunkname = '{}_Chunk{}'.format(info['name'], index)
+                    chunks[subd].append(chunkname)
+                    os.rename('Output', '../'+chunkname)
+                    os.remove(tgz)
           os.chdir(destpath)
           shutil.rmtree(subd)
      os.chdir(oldpath)
+     return 0
